Change interface: isolate procrustes.py, add vt_to_v attribute in create_UV_position_maps.py
